Remove outdated fields, collection methods, change collection interface, update postman tests
