+++ platform/ethereum/collection.go
@@ -14,114 +14,63 @@ func (p *Platform) GetCollections(owner string) (blockatlas.CollectionPage, erro
 	if err != nil {
 		return nil, err
 	}
-	page := NormalizeCollectionPage(collections, p.CoinIndex, owner)
-	return page, nil
+	return NormalizeCollections(collections, p.CoinIndex, owner), nil
 }
 
 func (p *Platform) GetCollectibles(owner, collectibleID string) (blockatlas.CollectiblePage, error) {
-	collection, items, err := p.collectionsClient.GetCollectibles(owner, collectibleID)
+	items, err := p.collectionsClient.GetCollectibles(owner, collectibleID)
 	if err != nil {
 		return nil, err
 	}
-	page := NormalizeCollectiblePage(collection, items, p.CoinIndex)
-	return page, nil
+	return NormalizeCollectiblePage(items, p.CoinIndex), nil
 }
 
-func NormalizeCollectionPage(collections []Collection, coinIndex uint, owner string) (page blockatlas.CollectionPage) {
+func NormalizeCollections(collections []Collection, coinIndex uint, owner string) (page blockatlas.CollectionPage) {
 	for _, collection := range collections {
-		if len(collection.Contracts) == 0 {
-			continue
-		}
 		item := NormalizeCollection(collection, coinIndex, owner)
-		if _, ok := supportedTypes[item.Type]; !ok {
-			continue
-		}
 		page = append(page, item)
 	}
-	return
+	return page
 }
 
 func NormalizeCollection(c Collection, coinIndex uint, owner string) blockatlas.Collection {
-	normalizeSupportedContracts(&c)
-	if len(c.Contracts) == 0 {
-		return blockatlas.Collection{}
-	}
-
-	description := blockatlas.GetValidParameter(c.Description, c.Contracts[0].Description)
-	symbol := blockatlas.GetValidParameter(c.Contracts[0].Symbol, "")
-	version := blockatlas.GetValidParameter(c.Contracts[0].NftVersion, "")
-	collectionType := blockatlas.GetValidParameter(c.Contracts[0].Type, "")
-
 	return blockatlas.Collection{
-		Name:            c.Name,
-		Symbol:          symbol,
-		Slug:            c.Slug,
-		ImageUrl:        c.ImageUrl,
-		Description:     description,
-		ExternalLink:    c.ExternalUrl,
-		Total:           int(c.Total.Int64()),
-		Id:              c.Slug,
-		CategoryAddress: c.Slug,
-		Address:         owner,
-		Version:         version,
-		Coin:            coinIndex,
-		Type:            collectionType,
+		Name:         c.Name,
+		ImageUrl:     c.ImageUrl,
+		Description:  c.Description,
+		ExternalLink: c.ExternalUrl,
+		Total:        int(c.Total.Int64()),
+		Id:           c.Slug,
+		Address:      owner,
+		Coin:         coinIndex,
 	}
 }
 
-func normalizeSupportedContracts(c *Collection) {
-	supportedContracts := make([]PrimaryAssetContract, 0)
-	for _, contract := range c.Contracts {
-		if _, ok := supportedTypes[contract.Type]; !ok {
-			continue
+func NormalizeCollectiblePage(collectibles []Collectible, coinIndex uint) (page blockatlas.CollectiblePage) {
+	for _, collectible := range collectibles {
+		item := NormalizeCollectible(collectible, coinIndex)
+		if _, ok := supportedTypes[item.Type]; ok {
+			page = append(page, item)
 		}
-		supportedContracts = append(supportedContracts, contract)
 	}
-	c.Contracts = supportedContracts
+	return page
 }
 
-func NormalizeCollectiblePage(c *Collection, srcPage []Collectible, coinIndex uint) (page blockatlas.CollectiblePage) {
-	normalizeSupportedContracts(c)
-	if len(c.Contracts) == 0 {
-		return
-	}
-	for _, src := range srcPage {
-		item := NormalizeCollectible(c, src, coinIndex)
-		if _, ok := supportedTypes[item.Type]; !ok {
-			continue
-		}
-		page = append(page, item)
-	}
-	return
-}
-
-func NormalizeCollectible(c *Collection, a Collectible, coinIndex uint) blockatlas.Collectible {
-	address := blockatlas.GetValidParameter(c.Contracts[0].Address, "")
-	collectionType := blockatlas.GetValidParameter(c.Contracts[0].Type, "")
-	externalLink := blockatlas.GetValidParameter(a.ExternalLink, a.AssetContract.ExternalLink)
+func NormalizeCollectible(a Collectible, coinIndex uint) blockatlas.Collectible {
 	id := strings.Join([]string{a.AssetContract.Address, a.TokenId}, "-")
 	return blockatlas.Collectible{
-		ID:               id,
-		CollectionID:     c.Slug,
-		ContractAddress:  address,
-		TokenID:          a.TokenId,
-		CategoryContract: a.AssetContract.Address,
-		Name:             a.Name,
-		Category:         c.Name,
-		ImageUrl:         a.ImagePreviewUrl,
-		ProviderLink:     a.Permalink,
-		ExternalLink:     externalLink,
-		Type:             collectionType,
-		Description:      a.Description,
-		Coin:             coinIndex,
+		ID:              id,
+		CollectionID:    a.Collection.Slug,
+		TokenID:         a.TokenId,
+		ContractAddress: a.AssetContract.Address,
+		Name:            a.Name,
+		Category:        a.Collection.Name,
+		ImageUrl:        a.ImagePreviewUrl,
+		ProviderLink:    a.Permalink,
+		ExternalLink:    a.Collection.ExternalLink,
+		Type:            a.AssetContract.Type,
+		Description:     a.Description,
+		Coin:            coinIndex,
+		Version:         a.AssetContract.Version,
 	}
-}
-
-func searchCollection(collections []Collection, collectibleID string) *Collection {
-	for _, i := range collections {
-		if strings.EqualFold(i.Slug, collectibleID) {
-			return &i
-		}
-	}
-	return nil
 }
