+++ platform/ethereum/collection_client.go
@@ -2,7 +2,6 @@ package ethereum
 
 import (
 	"github.com/trustwallet/blockatlas/pkg/blockatlas"
-	"github.com/trustwallet/blockatlas/pkg/errors"
 	"net/url"
 	"strconv"
 )
@@ -20,30 +19,7 @@ func (c CollectionsClient) GetCollections(owner string) (page []Collection, err 
 	return
 }
 
-func (c CollectionsClient) GetCollectibles(owner string, collectibleID string) (*Collection, []Collectible, error) {
-	collections, err := c.GetCollections(owner)
-	if err != nil {
-		return nil, nil, err
-	}
-	collection := searchCollection(collections, collectibleID)
-	if collection == nil {
-		return nil, nil, errors.E("collectible not found", errors.TypePlatformClient,
-			errors.Params{"collectibleID": collectibleID})
-	}
-
-	query := url.Values{
-		"owner": {owner},
-		"limit": {strconv.Itoa(300)},
-	}
-
-	query.Set("collection", collection.Slug)
-
-	var page CollectiblePage
-	err = c.Get(&page, "api/v1/assets", query)
-	return collection, page.Collectibles, err
-}
-
-func (c CollectionsClient) GetCollectiblesV4(owner string, collectibleID string) ([]Collectible, error) {
+func (c CollectionsClient) GetCollectibles(owner string, collectibleID string) ([]Collectible, error) {
 	query := url.Values{
 		"owner":      {owner},
 		"collection": {collectibleID},
